+++ server/interface/Admin.interface.ts
@@ -1,0 +1,10 @@
+export interface loginBody {
+  email: string;
+  password: string;
+}
+
+export interface registerBody {
+  name: string;
+  email: string;
+  password: string;
+}