+++ server/src/routes/admin.route.ts
@@ -1,8 +1,9 @@
 import express from "express";
 
-import { registerAdmin } from "../controllers/admin.controller";
+import { registerAdmin, loginAdmin } from "../controllers/admin.controller";
 
 const router = express.Router();
 router.post("/register", registerAdmin);
+router.post("/login", loginAdmin);
 
 export default router;